[ MVVM ] Rename android arch related utils classes
