+++ app/src/main/java/arch/Objects.java
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-package architecture_components.utils;
+package arch;
 
-/**
- * Status of a resource that is provided to the UI.
- * <p>
- * These are usually created by the Repository classes where they return
- * {@code LiveData<Resource<T>>} to pass back the latest data to the UI with its fetch status.
- *
- * https://github.com/googlesamples/android-architecture-components/blob/master/GithubBrowserSample/app/src/main/java/com/android/example/github/vo/Status.java
- */
-public enum Status {
-    SUCCESS,
-    ERROR,
-    LOADING
+class Objects {
+    public static boolean equals(Object o1, Object o2) {
+        if (o1 == null) {
+            return o2 == null;
+        }
+        if (o2 == null) {
+            return false;
+        }
+        return o1.equals(o2);
+    }
 }
